# Conflicts:
#	Modules/User/Routes/admin.php

+++ Modules/Brand/Http/Controllers/Admin/BrandController.php
@@ -19,7 +19,6 @@ class BrandController extends Controller
         protected BrandStorage $brandStorage,
 
     ) {
-        $this->authorizeResource(Brand::class, 'brand');
     }
 
     public function all()
@@ -33,6 +32,8 @@ class BrandController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Brand::class);
+
         $brands = $this->brandRepository->jsonPaginate();
 
         return BrandResource::collection($brands);
@@ -42,11 +43,15 @@ class BrandController extends Controller
     {
         $brand = $this->brandRepository->find($brand);
 
+        $this->authorize('view', $brand);
+
         return new BrandResource($brand);
     }
 
     public function store(BrandCreateRequest $request)
     {
+        $this->authorize('create',  Brand::class);
+
         $brandDto = BrandDto::fromFormRequest($request);
 
         $brand = $this->brandStorage->store($brandDto);
@@ -58,6 +63,8 @@ class BrandController extends Controller
     {
         $brand = $this->brandRepository->find($brand);
 
+        $this->authorize('update', $brand);
+
         $brand = $this->brandStorage->update(
             $brand, BrandDto::fromFormRequest($request)
         );
@@ -68,6 +75,8 @@ class BrandController extends Controller
     public function destroy(int $brand)
     {
         $brand = $this->brandRepository->find($brand);
+
+        $this->authorize('delete', $brand);
 
         $this->brandStorage->delete($brand);
